fix: polling loop logic
The polling loop logic was decrementing the counters and marking devices
as away before checking if they were online. So, if `offline_after` was
set to 1, instead of devices surviving until next poll when disconnected
the loop was marking all devices as away and then immediatelly marking
them online again.

This commit fixes the logic and achieves the intended behaviour by:
1. checking if devices are online *before* decrementing the counter
2. only decrementing the counter for offline devices